requirements.txt
Django~=4.2.7
django-cors-headers
pillow~=10.3.0
djangorestframework~=3.15.2
djangorestframework-simplejwt
drf-spectacular